Add final score to played games
