+++ components/Game/GoalieComparison.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import styled from 'styled-components';
-import { GoalieStats, Matchup } from '../../../pages/api/v1/schedule/game/[gameId]';
-import { ComparisonHeader, FlexColumn, FlexRow, SectionTitle, TeamLogoSmall } from '../common';
+import { GoalieStats, Matchup } from '../../pages/api/v1/schedule/game/[gameId]';
+import { ComparisonHeader, FlexColumn, FlexRow, SectionTitle, TeamLogoSmall } from './common';
 
 interface Props {
   gameData: Matchup;
