+++ components/Game/TeamStats.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import styled from 'styled-components';
-import { Matchup } from '../../../pages/api/v1/schedule/game/[gameId]';
-import { FlexRow, SectionTitle, TeamLogoSmall } from '../common';
+import { Matchup } from '../../pages/api/v1/schedule/game/[gameId]';
+import { FlexRow, SectionTitle, TeamLogoSmall } from './common';
 
 interface Props {
   gameData: Matchup;
